Refactor: transform files
