+++ internal/export/dashboards.go
@@ -18,18 +18,6 @@ import (
 	"github.com/elastic/elastic-package/internal/packages"
 )
 
-var (
-	encodedFields = []string{
-		"attributes.kibanaSavedObjectMeta.searchSourceJSON",
-		"attributes.layerListJSON",
-		"attributes.mapStateJSON",
-		"attributes.optionsJSON",
-		"attributes.panelsJSON",
-		"attributes.uiStateJSON",
-		"attributes.visState",
-	}
-)
-
 // Dashboards method exports selected dashboards with references objects. All Kibana objects are saved to local files
 // in appropriate directories.
 func Dashboards(kibanaDashboardsClient *dashboards.Client, dashboardsIDs []string) error {
